requirements/local.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/production.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.
-r production.txt

# Deployment
Fabric

# Documentation
Sphinx

# Testing
nose
mock
django-nose
coverage
pep8
pyflakes
pylint
scrapy
six
django-allauth